refactor(rust): standardize compression detection
