+++ crates/polars-io/src/utils/other.rs
@@ -6,7 +6,6 @@ use once_cell::sync::Lazy;
 use polars_core::prelude::*;
 #[cfg(any(feature = "ipc_streaming", feature = "parquet"))]
 use polars_core::utils::{accumulate_dataframes_vertical_unchecked, split_df_as_ref};
-use polars_error::to_compute_err;
 use polars_utils::mmap::MMapSemaphore;
 use regex::{Regex, RegexBuilder};
 
@@ -43,46 +42,6 @@ pub fn get_reader_bytes<'a, R: Read + MmapBytesReader + ?Sized>(
             reader.read_to_end(&mut bytes)?;
             Ok(ReaderBytes::Owned(bytes))
         }
-    }
-}
-
-/// Decompress `bytes` if compression is detected, otherwise simply return it.
-/// An `out` vec must be given for ownership of the decompressed data.
-pub fn maybe_decompress_bytes<'a>(bytes: &'a [u8], out: &'a mut Vec<u8>) -> PolarsResult<&'a [u8]> {
-    assert!(out.is_empty());
-    use crate::prelude::is_compressed;
-    let is_compressed = bytes.len() >= 4 && is_compressed(bytes);
-
-    if is_compressed {
-        #[cfg(any(feature = "decompress", feature = "decompress-fast"))]
-        {
-            use crate::utils::compression::magic::*;
-
-            if bytes.starts_with(&GZIP) {
-                flate2::read::MultiGzDecoder::new(bytes)
-                    .read_to_end(out)
-                    .map_err(to_compute_err)?;
-            } else if bytes.starts_with(&ZLIB0)
-                || bytes.starts_with(&ZLIB1)
-                || bytes.starts_with(&ZLIB2)
-            {
-                flate2::read::ZlibDecoder::new(bytes)
-                    .read_to_end(out)
-                    .map_err(to_compute_err)?;
-            } else if bytes.starts_with(&ZSTD) {
-                zstd::Decoder::new(bytes)?.read_to_end(out)?;
-            } else {
-                polars_bail!(ComputeError: "unimplemented compression format")
-            }
-
-            Ok(out)
-        }
-        #[cfg(not(any(feature = "decompress", feature = "decompress-fast")))]
-        {
-            panic!("cannot decompress without 'decompress' or 'decompress-fast' feature")
-        }
-    } else {
-        Ok(bytes)
     }
 }
