+++ crates/polars-mem-engine/src/executors/scan/csv.rs
@@ -4,6 +4,7 @@ use polars_core::config;
 use polars_core::utils::{
     accumulate_dataframes_vertical, accumulate_dataframes_vertical_unchecked,
 };
+use polars_io::utils::compression::maybe_decompress_bytes;
 
 use super::*;
