+++ crates/polars-plan/src/plans/conversion/scans.rs
@@ -3,6 +3,7 @@ use polars_io::path_utils::is_cloud_url;
 #[cfg(feature = "cloud")]
 use polars_io::pl_async::get_runtime;
 use polars_io::prelude::*;
+use polars_io::utils::compression::maybe_decompress_bytes;
 use polars_io::RowIndex;
 
 use super::*;
